+++ frontend/components/chat/Model.tsx
@@ -4,9 +4,8 @@ import { useCallback } from "react";
 import Select from "../general/selects/Select";
 import { Input } from "../ui";
 import { Slider } from "../ui/slider";
-import { RotateCcw } from "lucide-react";
-import MenuButton from "../general/buttons/MenuButton";
-import useModelStore, { ConfigType } from "@/store/model";
+import useModelStore from "@/store/model";
+import ModelResetButton from "../buttons/ModelResetButton";
 
 export default function Model() {
   const [model, config, setModelStore] = useModelStore((state) => [
@@ -22,33 +21,13 @@ export default function Model() {
     },
     [config, setModelStore]
   );
-  const onClickReset = useCallback(() => {
-    if (!model) return;
-
-    const newConfig: ConfigType = {};
-    model.configs.forEach((c) => {
-      if (!c.default) {
-        newConfig[c.name] = "";
-      } else {
-        newConfig[c.name] = c.default;
-      }
-    });
-
-    setModelStore({ config: newConfig });
-  }, [model, setModelStore]);
 
   if (!model) return null;
 
   return (
     <div className="w-full">
-      <div className="sticky top-0 md:top-14 pl-[6.75rem] md:pl-0 md:ml-3 w-full pb-3 md:pb-0 md:h-0 z-20 bg-background bg-background">
-        <MenuButton
-          className="w-24 md:w-28"
-          Icon={RotateCcw}
-          text="Reset"
-          onClick={onClickReset}
-          responsive
-        />
+      <div className="sticky top-0 md:top-14 pl-[6.75rem] md:pl-0 w-full pb-3 md:pb-0 md:h-0 z-20 bg-background">
+        <ModelResetButton />
       </div>
       <div className="w-full flex flex-col gap-7 md:pl-[8.5rem] md:py-3 whitespace-pre-wrap text-sm">
         <div className="text-base font-bold text-theme-foreground bg-theme px-2 py-1 rounded-md">
